Add files via upload
First, create a Virtual Environment and install the packages and libraries required using the given requirements.txt. Activate the virtual environment. Create a django project with name face_identification, replace the entire project contents using the given contents "except the settings.py file in the face_identification folder". For the same file in your django project, copy the contents that are present extra in the settings.py provided in this repository. If everything has been performed correctly, the application works as intended.